Implement type successor
Depending on your use case you might want to define some successor who takes over all relations that were owned by the entity that is going to be deleted.

Implementation assumes that the entity to be delelted should hold the information about its successor.

With type SUCCESSOR and by adding annotation onSoftDeleteSuccessor to one of the properties of softdeleted entity, all relations are replaced to point to the new entity instead.